+++ sw.js
@@ -1,9 +1,9 @@
-const CACHE_NAME = "spoolmate-v118";
+const CACHE_NAME = "spoolmate-v119";
 const APP_SHELL = [
   "./",
   "./index.html",
-  "./styles.css?v=170",
-  "./app.js?v=170",
+  "./styles.css?v=171",
+  "./app.js?v=171",
   "./manifest.webmanifest",
   "./icons/icon.svg",
   "./icons/icon-192.png",